devices: vhost-user: GPU: Abort tasks at destruction
The GPU device spawns abortable tasks, and depends on them being
properly destroyed. For example, gfxstream is never shut down if
GPU Frontend being held by the run_display task isn't freed.

This aborts the running tasks at the time of destruction. That
causes the aborted task to be put back on the executor's queue,
so we run it for one more step to fully deallocate the task.

BUG=b:260248096
TEST=downstream & presubmit

Change-Id: I3f98f520082273682d0afd7ab928aaa9b8c19197
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4010952
Commit-Queue: Idan Raiter <idanr@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>