Allow exceptions to needless return
This CL adds allow directives for needless return on functions that have
void returns and use the pattern of "error and then return early", and
happen to have a currently-needless return as the last statement. This
is to prevent future bugs if the functions are extended.

Each case is documented with a comment explaining why the exception is
allowed.

BUG=chromium:908640, b:157245930
TEST=cargo clippy with needless_return suppression removed

Change-Id: I44f978b0cd9aa0d574acc73cd2f0a34f1b181e84
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4018392
Auto-Submit: Clarissa Garvey <clarissagarvey@chromium.org>
Commit-Queue: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Dennis Kempin <denniskempin@google.com>