crosvm: Add a parameter to set hybrid CPU type in "--cpu"
Add a parameter "core-types" in "--cpu" option to set hybrid type for
vCPUs. The valid hybrid types include "atom" and "core" for x86.

For x86, with this option, guest can get the hybrid type of its vCPU via
CPUID.

This feature depends on the underlying hardware's capability. If pCPUs
are not hybrid architecture, this option cannot be set.

Co-Developed-by: Zhuocheng Ding <zhuocheng.ding@intel.com>
Co-Developed-by: Alexandre Courbot <acourbot@chromium.org>

BUG=b:255908920
TEST=presubmit and set `--cpus 8,core-types=[atom=[1,3-7],core=[0,2]]`

Change-Id: Ifcf50e1cb3d8083c9b5c531607a80670aa39b3e6
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4020039
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>