devices: virtio: fix a memory ordering bug in Queue::peek on arm
Currently, `Queue::peek` does the following steps for reading and
checking `avail_index`. 1. Reads the current `avail_index` 2. Checks if
`avail_index` meets the condition to go further. 3. If so, read the
current the descriptor index. This works fine on x86. However, on Arm,
which has weak-memory model, the memory read for the third step could
happen before the memory read for the first step since there is no fence
between the first and the third steps. That causes reading a stale value
for the descriptor.

This commit adds a fence between them to fix potential race condition
bugs caused by reading stale descriptor values.

Note it is not fine to move the fence in `Queue::get_avail_index` from
the beginning of it to the end of `get_avail_index`. This is because all
other virtqueue access APIs such as `set_avail_event` also have fences
at the beginning of them. If you move the fence in `get_avail_index` to
the end of `get_avail_index`, the sequence of `set_avail_event` and then
`get_avail_index` loses a fence between them.  Because this is a "write
-> read" operations, losing a fence between them causes memory order bug
in both x86 and arm. Thus, we add another fence in this commit instead.

BUG=b:258630435
TEST=`./tools/dev_container ./tools/presubmit` passed

Change-Id: I693f1bc511fcc6fc2f3b0d7739c318483992280f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4041909
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Takaya Saeki <takayas@chromium.org>