crosvm: use structure to specify CPU options
We currently have many different CPU-related options. This does not play
well with the idea of having them in a configuration file, where they
would ideally be grouped under a single "cpus" section.

Turn the `cpus` option into a struct, which currently only has one
member specifying the number of cores we want. This makes the struct
compatible with the existing argument (since the first argument of a
struct does not need to be explicitly mentioned), which allowing us to
extend the struct with more options in the future.

This allows us to enable the `cpus` option in configuration files.

BUG=b:255223604
TEST=cargo test parse_cpu_opts

Change-Id: I1bc856a935be4774926dcba8b0714f5835feeed9
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4033212
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>