crosvm: use structure to specify memory options
We currently have many different memory-related options. This does not
play well with the idea of having them in a configuration file, where
they would ideally be grouped under a single "mem" section.

Turn the `mem` option into a struct, which currently only has one member
specifying the desired RAM size. This makes the struct compatible with
the existing argument (since the first argument of a struct does not
need to be explicitly mentioned), which allowing us to extend the struct
with more options in the future.

This allows us to enable the `mem` option in configuration files.

BUG=b:255223604
TEST=cargo test parse_mem_opts

Change-Id: I24c324002e4e399e5be492bbc9ddcac7fd2b349e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4033213
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>