devices: virtio: queue: remove avail_len check
The `Queue::peek()` function calculated `avail_len`, which is the number
of available ring entries between the device's most recently popped
available ring entry (`self.next_avail`) and the driver's most recently
published available ring entry (`avail_index`). This `avail_len` was
then compared against the configured queue size, and if `avail_len` was
larger than the queue size, the queue was assumed to be empty.

This check does not serve any purpose; we only want to check whether
there is at least one un-popped available ring entry, which is true if
and only if `self.next_avail` is not equal to the driver's `avail_index`
field. The `self.next_avail == avail_index` check is sufficient to
determine whether the queue is empty or not.

BUG=b:259992959
TEST=Boot Linux on crosvm on x86_64 and arm platforms
TEST=cargo test -p devices virtio::queue

Change-Id: Iae6ecf658fa4ea5c4d510274dac2f319317dd640
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4043790
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Noah Gold <nkgold@google.com>