devices: virtio: queue: enforce size invariants in set_size()
The virtio specification says that the driver must not write invalid
values (0, non power of two, etc.) to the queue_size field, so we can
disregard out-of-spec writes from the driver that do not meet these
requirements.

Our Queue used to validate these each time a descriptor chain was
popped, but it is simpler to validate the requirements during setup
when the driver is modifying the queue_size field.

Change the Queue to ignore writes that are not valid according to the
specification so that the `size` field is always valid and does not need
to be validated during `peek()`/`pop()`.

This also means that the `actual_size()` function is no longer
necessary, since `size()` will always return the actual size.

BUG=None
TEST=Boot x86-64 Linux in crosvm

Change-Id: I90be114e0bbfbe5d3e3aa038c235fccde4c00271
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4045081
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>