devices: virtio: queue: use masking to wrap queue index
Since the queue size must be a power of two for split virtqueues, we can
use a simpler bit masking operation to calculate the modulus rather than
the `%` operator, which required an actual division operation since the
divisor was not known at compile time.

This probably makes very little actual difference in performance, but we
might as well take advantage of it anyway.

BUG=None
TEST=Boot x86-64 Linux in crosvm
TEST=Verify that no div/idiv appears in peek disassembly

Change-Id: I905e69c1f7c58db1d3a4ae94e72e7a6ff513cee4
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4045082
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>