serde_keyvalues: perform integer detection up to separator character
Integer detection was stopping and succeeding at the first non-integer
character, i.e. "0-4" would be detected as integer "0" and leave "-4" as
the remainder string, meaning the parser would fail later.

Improve this by requesting integer detection to also detect a separator
character and fail if the whole input until that separator is not a
valid number. That way, "0-4" fails to parse as a number, and
deserialize_any will instead properly recognize it as a string.

BUG=b:255223604
TEST=cargo test -p serde_keyvalues

Change-Id: I6b5805c7b0b82c36f61280c26b9f92f219196511
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4060599
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>