devices: virtio: vhost-user: move executor creation before jail.
This commit has  vhost-user devices implemented by `VhostUserDevice`
trait call `Executor::new()` before jailing, not after jailing . The
context is the jail for io_uring. For the effective io_uring jail, an
`Executor` must be created before jailing.

Currently, the virtio-block and the virtio-console are implemented by
`VhostUserDevice` trait. Thus, this change enables vhost-user block
and console to work with io_uring in the sandbox.

This change also updates the doc comment of `VhostUserDevice`.
`Executor` can be actually kept during jailing as this change does, so
the current doc comment is obsoleted. However, `VhostUserDevice` is
still powerful to ensure that every initialization for
`VhostUserBackend` is done in the jailed process.

BUG=b:251039942
BUG=b:262341514
TEST=`crosvm devices --async_executor uring --block ...` works with the frontend.
TEST=`crosvm devices --async-executor uring --jail --block ...` works with the frontend.
TEST=`crosvm devices --async_executor uring --serial ...` works with the frontend.
TEST=`crosvm devices --async-executor uring --jail --serial ...` works with the frontend.

Change-Id: I1365ef96c2f2f27cd63bdeb3d0ca327100a5dcdc
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4098934
Commit-Queue: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>