devices: gpu: reduce size of GpuResponse with Box
Before this change, EdidBytes was the largest member, and it caused the
VirtioGpuResult type to be 136 bytes. Most cases do not need this
member, so store it in a Box to dynamically allocate space for EdidBytes
only when needed.

Fixes a new clippy result_large_err lint in Rust 1.65.

BUG=b:260784028
TEST=tools/clippy

Change-Id: I8eb81cc184ff98a4c4396b8828e7192b4c6e8ccc
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4064710
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>