crosvm: add CPU clusters to "--cpu" option
Add a "clusters" parameter to the "--cpu" options that allows to specify
CPU clusters as a sequence of CpuSets. This enables us to concentrate
more CPU configuration under the same parameter and looks better in the
configuration file.

The old "--cpu-cluster" parameter is kept for compatiblity, but
deprecated.

BUG=b:255223604
TEST=cargo test

Change-Id: Ic5b8aa7527c373d0797e20b4d88d8861488e0437
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4060287
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>