crosvm: create jails before their devices
We want to make the device creation methods consume their builder, but
a reference to the builder is also needed in order to create the jail.
Use the new virtio transport type to ensure that jails are always create
before and that the device creation is the final use of the builder.

BUG=b:179756087
TEST=cargo test

Change-Id: Ia39014351a41a4736c88ea531150e4324bf1e35e
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4075546
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Takaya Saeki <takayas@chromium.org>