crosvm: consume VirtioDeviceBuilder when creating devices
Creating a virtio device can mean consuming non-reusable resources, like
a tube. For this reason, making the device creation methods of
VirtioDeviceBuilder work with a reference on self is dangerous, as the
resource contained in the builder will need to be taken using some kind
of interior mutability, and the method cannot be called again safely
since the resource won't be present anymore.

Now that we are able to build jails before devices, make
`create_virtio_device` and `create_vhost_user_device` consume `self`.
This ensures that devices requiring Tubes are only created once, and
devices for which the builder can be reused just need to imlpement
VirtioDeviceBuilder on a reference of themselves.

BUG=b:179756087
TEST=cargo test

Change-Id: Ie9d7f71bd7f61c4160d40b3aeec881df7cca492d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4075547
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>