devices: virtio: return a Result from activate()
If a device fails to activate, it can now return an error, which will
cause the device to be marked as DEVICE_NEEDS_RESET so the driver can
tell that initialization failed.

This allows devices to report an error during activation more naturally.
Many devices had error logging that just left the device in an
inoperable state without reporting it to the driver.

Most devices leave their internal variables in an inconsistent state
when activate() fails, so this will not necessarily allow the driver to
successfully reset the device, but this is a first step toward making it
possible.

BUG=None
TEST=tools/presubmit --all

Change-Id: Id49ec1c3f19384c2585235ba4f9cf76ee4253641
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4081631
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>