hypervisor: add default enable_capability impl
Since the newly-added Vm::enable_capability() function is only required
for hypervisors which support VmCap::BusLockDetect, we can add a default
implementation of the function in the trait for other hypervisors. This
replaces the version from HaxmVm and also implements it for WhpxVm,
which is not enabled by default.

Additionally, handle the VmCap::BusLockDetect case in the WpxVm
check_capability() function to fix compilation.

BUG=None
TEST=Build for Windows with whpx feature enabled

Fixes: 59f6611a74b3 ("Add bus lock detect support in CROSVM")
Change-Id: I8a4ed7c14a224439dbba79d44134219658786e8f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4087466
Reviewed-by: Clarissa Garvey <clarissagarvey@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Yiming Zhang <yiming.zhang@intel.corp-partner.google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>