broker: Actually pass ui_exceptions to sandbox
We were always passing a `0` for `ui_exceptions` for the sandbox
policy's job object, where we should have been passing the value
specified in the policy. This meant we were stricter than we specified
in the policy, but also that we could not be as strict with our job
object as we wanted to be, as we couldn't make use of exceptions.

BUG: b:262445398
TEST: Built and ran windows crosvm downstream.
Change-Id: I7df0282dbfadb33550956c2461b7dc084116617f
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4103909
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Commit-Queue: Richard Otap <rotap@google.com>