devices: vhost-user device: remove per-backend max_vring_len()
There is no mechanism in the vhost-user protocol for the VMM side to
ask the backend device for its max queue size, so the only way to ensure
compatibility with arbitrary VMM frontends is for the backend to support
the maximum valid size. The maximum queue size for both packed and split
virtqueues is 32768 (2**15), so we can use that as the max queue size
for all device types. The VMM frontend can still choose a smaller queue
size when it sends the VHOST_USER_SET_VRING_NUM request.

BUG=b:262291811
TEST=tools/presubmit --all
TEST=Boot x86-64 Linux from a vhost-user block device attached to crosvm

Change-Id: Ie685ea7718f738ef19c1bc27132bbe5fc0466607
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4098493
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>