cmdline: fix crosvm run help string
The ///-style docstring at the top of RunCommand was included in the argh
help output (`crosvm --help` and `crosvm run --help`), but the bulk of
the comment is intended for developers, not for users of crosvm.

Move the run subcommand's help to an explicit #[argh(description = "")]
so the docstring is not used as the help text.

BUG=None
TEST=crosvm --help

Change-Id: Iccb702203f376c304fc2825eec42471a3b48efd9
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4109471
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>