devices: virtio: fix missing deduplication of keep_rds
Minijail requires the `keep_rds` list to be unique, so you need to
deduplicate it before passing it to Minijail. This commit fixes missing
deduplication for the vhost-user devices.

This commit also fixes the wrong usage of `Vector::dedup()` of the
vhost-user fs device.

BUG=b:262339813
TEST=./tools/presubmit passes

Change-Id: Id8d8a5fb327a75a938315601ba67a91a698972c6
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4112720
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Takaya Saeki <takayas@chromium.org>