cmdline: parse --vfio with serde_keyvalue
The only incompatible change to the command line is dropping support for
pci-address=auto, which is equivalent to not specifying pci-address.
This option was not actually used in practice in crosvm-direct.sh.

Also enable the vfio option in the config file. The vfio-platform option
is deprecated, so it is still skipped in config serialization.

BUG=b:255223604
TEST=cargo test
TEST=cargo test --features=direct

Change-Id: I35ad8e5fbdc18ad1056dba50957cd1c724675d74
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4116059
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>