devices: fs: hold inodes lock throughout add_entry()
There seems to be no downside to holding self.inodes.lock() for the
whole PassthroughFs::add_entry() function. We don't do any siginificant
amount of work while the lock is held; either the inode already exists,
in which case we increment its refcount, or the inode is added to the
map. Note that we already opened the file before entering this function,
so the lock is not held across open().

This also prevents the possible race between threads opening the same
inode simultaneously, so remove the related comment.

BUG=b:261922849
TEST=boot arcvm on brya

Change-Id: I036ebaed179f7ac43d071a26cb68ba9a131521a0
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4116247
Reviewed-by: Morg <morg@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>