media: cros-codecs: stop exposing backend to the user
The backend object should be an implementation detail and not visible to
users. Replace the methods allowing to access the backend from outside
of the crate by new methods doing what we wanted to do with it, and add
new per-backend constructor methods so users don't need to create the
backend themselves.

This allows us to simplify the backend interface a bit, and also to make
the backend traits and structs crate-private.

BUG=b:214478588
TEST=cargo test --features vaapi -p cros-codecs

Change-Id: Ifcd144176f21016d84dd8f39fee57e4eb48ca87b
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4120374
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>