media: cros-codecs: remove the `stateless` module
All decoders and backends in cros-codecs are stateless at the moment, so
it is a bit strange and confusing to have a `stateless` module under the
already stateless decoder module.

Let's remove it for now and simplify our module hierarchy a bit as it
was easy (for me at least) to get lost.

BUG=b:214478588
TEST=cargo test --features vaapi -p cros-codecs

Change-Id: Id6925453caab28d47a58aea856a6ee6ebf521ed9
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4120373
Reviewed-by: Daniel Almeida <daniel.almeida@collabora.corp-partner.google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>