media: cros-codecs: vaapi: remove image_formats field
Querying the supported image formats is an operation that happens rarely
(when starting a stream or changing the resolution) and that may depend
on the dynamic conditions of the stream itself. Therefore there is
little benefit or incentive to cache it once and for all when the
backend it created - just do it on demand and simplify the backend
structs a bit.

BUG=b:214478588
TEST=cargo test --features vaapi -p cros-codecs

Change-Id: Ie5507577a0e0bd527b43955e1c551e6cb1d6a09d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4120376
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Daniel Almeida <daniel.almeida@collabora.corp-partner.google.com>