media: cros-codecs: remove dependency on downcast_rs
We only use this for testing, and this can be replaced at low cost by a
method returning a &dyn Any and having the tests perform the cast on it.
Since downcast_rs gets in the way of some of the refactoring we want to
do, let's remove it.

BUG=b:214478588
TEST=cargo test --features vaapi -p cros-codecs

Change-Id: I343e0a8e0f219f6bcf3cbc5efac82a1ed999a9a0
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4123651
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>